+++ app/routes/index.tsx
@@ -6,7 +6,7 @@ import { Pagination } from '../components/Pagination'
 import { PageHeading } from '../components/PageHeading'
 import type { Meta } from '../types/meta'
 import { config } from '../siteSettings'
-import { SearchForm } from '../islands/SearchForm'
+import { SearchForm } from '../components/SearchForm'
 
 export default createRoute(async (c) => {
   const client = getMicroCMSClient(c)
@@ -45,7 +45,7 @@ export default createRoute(async (c) => {
       {/* 訪問記録一覧 */}
       <div class="space-y-6">
         <PageHeading>訪問記録</PageHeading>
-        <SearchForm redirectToSearchPage={true} placeholder="訪問記録を検索..." />
+        <SearchForm initialQuery="" />
         {totalCount === 0 ? (
           <p class="text-gray-500">まだ訪問記録がありません</p>
         ) : (
